requirements.txt
Louie
suds
lxml
aiohttp
aiofiles
pydispatcher
certifi
